requirements.txt
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
beautifulsoup4==0.0.1
requests==2.24.0
django-extensions==3.1.1
requests-html==0.10.0